backend/requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-picklefield==3.1
django-q==1.3.9
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
